requirements-lock.txt
absl-py==2.2.1
accelerate==1.13.0
alembic==1.18.4
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
barrel==0.3.0
beautifulsoup4==4.13.3
biopython==1.87
black==26.3.1
bleach==6.2.0
blinker==1.9.0
blis==1.2.0
boolean.py==5.0
branca==0.8.1
CacheControl==0.14.4
cached-property==1.5.2
cachetools==5.5.2
catalogue==2.0.10
category_encoders==2.8.1
certifi==2025.1.31
cffi==1.17.1
cfgv==3.5.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
colorlog==6.10.1
combine==2.17.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
Counter==1.0.0
coverage==7.13.5
cv==1.0.0
cycler==0.12.1
cyclonedx-python-lib==11.7.0
cymem==2.0.11
dash==3.0.1
dask==2025.3.0
dateparser==1.2.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
et_xmlfile==2.0.0
execnet==2.1.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.29.0
flake8==7.1.2
Flask==3.0.3
flatbuffers==25.2.10
folium==0.19.5
fonttools==4.56.0
fqdn==1.5.1
fsspec==2025.3.0
gast==0.6.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
hf-xet==1.4.3
honcho==1.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface_hub==1.11.0
identify==2.6.19
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.6.1
iniconfig==2.3.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaggle==1.7.4.5
keras==3.9.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
librt==0.9.0
license-expression==30.4.4
llvmlite==0.46.0
locket==1.0.0
Mako==1.3.11
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.2
murmurhash==1.0.12
mypy==1.20.1
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.32.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.8.1
nodeenv==1.10.0
notebook==7.3.3
notebook_shim==0.2.4
numba==0.64.0
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.1
optuna==4.8.0
overrides==7.7.0
packageurl-python==0.17.6
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pathspec==1.0.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.10.0
platformdirs==4.3.7
plotly==6.0.1
pluggy==1.6.0
pre_commit==4.5.1
preprocessing==0.1.13
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-serializable==2.1.0
py4j==0.10.9.7
pyarrow==19.0.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pyflakes==3.2.0
Pygments==2.19.1
pyparsing==3.2.3
pyspark==3.5.5
pytest==9.0.3
pytest-cov==7.1.0
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-discovery==1.2.2
python-frontmatter==1.1.0
python-json-logger==3.3.0
python-slugify==8.0.4
pytokens==0.4.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2026.4.4
repaint==0.1.0
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.24.0
ruff==0.15.11
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
shap==0.51.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sphinx-rtd-theme==0.2.4
SQLAlchemy==2.0.49
srsly==2.5.1
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.44.0
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
terminado==0.18.1
text-unidecode==1.3
textblob==0.19.0
thinc==8.3.4
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.2
toml==0.10.2
tomli==2.4.1
tomli_w==1.2.0
toolz==1.0.0
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==5.5.4
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.3.0
virtualenv==21.2.4
wasabi==1.1.3
watchfiles==0.19.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wordcloud==1.9.4
wrapt==1.17.2
xgboost==3.0.0
xlrd==1.2.0
xyzservices==2025.1.0
zipp==3.21.0
